Refactor: Extract new mod `async_read_utility::bytes_impl`
