+++ src/async_read_utility/bytes_impl.rs
@@ -1,0 +1,141 @@
+use super::{read_to_container_rng, Container, ReadToContainerRngFuture};
+
+use std::{
+    future::Future,
+    io::Result,
+    marker::Unpin,
+    mem::MaybeUninit,
+    pin::Pin,
+    task::{Context, Poll},
+};
+
+use bytes::{BufMut, BytesMut};
+use tokio::io::AsyncRead;
+
+#[cfg_attr(docsrs, doc(cfg(feature = "read-exact-to-bytes")))]
+impl Container for BytesMut {
+    fn reserve(&mut self, n: usize) {
+        BytesMut::reserve(self, n)
+    }
+
+    fn capacity(&self) -> usize {
+        BytesMut::capacity(self)
+    }
+
+    unsafe fn spare_mut(&mut self) -> &mut [MaybeUninit<u8>] {
+        self.chunk_mut().as_uninit_slice_mut()
+    }
+
+    unsafe fn advance(&mut self, n: usize) {
+        self.advance_mut(n)
+    }
+}
+
+/// Returned future of [`read_exact_to_vec`].
+#[derive(Debug)]
+#[cfg_attr(docsrs, doc(cfg(feature = "read-exact-to-bytes")))]
+pub struct ReadExactToBytesFuture<'a, R: ?Sized>(ReadToBytesRngFuture<'a, R>);
+
+impl<R: AsyncRead + ?Sized + Unpin> Future for ReadExactToBytesFuture<'_, R> {
+    type Output = Result<()>;
+
+    fn poll(mut self: Pin<&mut Self>, cx: &mut Context<'_>) -> Poll<Self::Output> {
+        Pin::new(&mut self.0).poll(cx)
+    }
+}
+
+#[cfg_attr(docsrs, doc(cfg(feature = "read-exact-to-bytes")))]
+/// * `nread` - bytes to read in
+///
+/// Return [`std::io::ErrorKind::UnexpectedEof`] on Eof.
+///
+/// NOTE that this function does not modify any existing data.
+///
+/// # Cancel safety
+///
+/// It is cancel safe and dropping the returned future will not stop the
+/// wakeup from happening.
+pub fn read_exact_to_bytes<'a, R: AsyncRead + ?Sized + Unpin>(
+    reader: &'a mut R,
+    bytes: &'a mut BytesMut,
+    nread: usize,
+) -> ReadExactToBytesFuture<'a, R> {
+    ReadExactToBytesFuture(read_to_bytes_rng(reader, bytes, nread..=nread))
+}
+
+/// Returned future of [`read_exact_to_vec`].
+#[derive(Debug)]
+#[cfg(feature = "read-exact-to-bytes")]
+#[cfg_attr(docsrs, doc(cfg(feature = "read-exact-to-bytes")))]
+pub struct ReadToBytesRngFuture<'a, Reader: ?Sized>(ReadToContainerRngFuture<'a, BytesMut, Reader>);
+
+#[cfg(feature = "read-exact-to-bytes")]
+impl<Reader: AsyncRead + ?Sized + Unpin> Future for ReadToBytesRngFuture<'_, Reader> {
+    type Output = Result<()>;
+
+    fn poll(mut self: Pin<&mut Self>, cx: &mut Context<'_>) -> Poll<Self::Output> {
+        Pin::new(&mut self.0).poll(cx)
+    }
+}
+
+#[cfg(feature = "read-exact-to-bytes")]
+#[cfg_attr(docsrs, doc(cfg(feature = "read-exact-to-bytes")))]
+/// * `rng` - The start of the range specify the minimum of bytes to read in,
+///           while the end of the range specify the maximum of bytes that
+///           can be read in.
+///           If the lower bound is not specified, it is default to 0.
+///           If the upper bound is not specified, it is default to the
+///           capacity of `bytes`.
+///           The lower bound must not be larger than the upper bound.
+///
+/// Return [`std::io::ErrorKind::UnexpectedEof`] on Eof.
+///
+/// NOTE that this function does not modify any existing data.
+///
+/// # Cancel safety
+///
+/// It is cancel safe and dropping the returned future will not stop the
+/// wakeup from happening.
+pub fn read_to_bytes_rng<'a, R: AsyncRead + ?Sized + Unpin>(
+    reader: &'a mut R,
+    bytes: &'a mut BytesMut,
+    rng: impl std::ops::RangeBounds<usize>,
+) -> ReadToBytesRngFuture<'a, R> {
+    ReadToBytesRngFuture(read_to_container_rng(reader, bytes, rng))
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use tokio::io::AsyncWriteExt;
+
+    #[test]
+    fn test_read_exact_to_bytes() {
+        tokio::runtime::Builder::new_current_thread()
+            .enable_all()
+            .build()
+            .unwrap()
+            .block_on(async {
+                let (mut r, mut w) = tokio_pipe::pipe().unwrap();
+
+                let w_task = tokio::spawn(async move {
+                    for n in 1..=255 {
+                        w.write_u8(n).await.unwrap();
+                    }
+                });
+
+                let r_task = tokio::spawn(async move {
+                    let mut buffer = bytes::BytesMut::new();
+                    buffer.put_u8(0);
+
+                    read_exact_to_bytes(&mut r, &mut buffer, 255).await.unwrap();
+
+                    for (i, each) in buffer.iter().enumerate() {
+                        assert_eq!(*each as usize, i);
+                    }
+                });
+                r_task.await.unwrap();
+                w_task.await.unwrap();
+            });
+    }
+}
